refactor: Refactor code to gain 100% test coverage
There was a conflict regarding strtotime between Psalm and PHPUnit:
Psalm marked strtotime as being able to return false, so I added an
Exception, which I could never trigger due to the previous checks
that (hopefully) make invalid dates impossible.

PHPUnit on the other hand wants the Exception to be thrown in at
least one test; so I moved it into a helper method strtotime which
throws an exception if strtotime returns false, which can be tested.

Refactoring in stringToArray was caused by style issues; personally, I
prefer to write down every option into a if statement, but Psalm will
admonish this as being unecessary. True, but I find it more readable.
Anyway, this leaves the Exception added for Psalm untestable, so Psalm
got its way.